ci:set image registry explicitly on image build script (#38)
This commit modify the image build script in order to set explicitly a image registry for both global and regional control plane server images.

That is considered a good practice because it avoids to send images to unwanted registries.

Also, it solves a compatibility issue caused by divergent default behaviors of different container engines when images are built without set a registry.